+++ src/cubature_hexa.cpp
@@ -89,7 +89,7 @@ void cubature_hexa::set_rule(int in_rule)
 
 // method to get number of cubature points
 
-double cubature_hexa::get_n_pts(void)
+int cubature_hexa::get_n_pts(void)
 {
   return n_pts;
 }
